Rename AppHealthComponent to reflect tooltip
