+++ src/js/components/AppHealthBarWithTooltipComponent.jsx
